dwarf: Fix default size check for declarations
The dwarf_bytesize() can return -1 if a type has no size info like for
declarations.  As Dwarf_Word is an unsigned type, we should check
minus values properly.

+++ utils/dwarf.c
@@ -379,7 +379,7 @@ static size_t type_size(Dwarf_Die *die, size_t default_size)
 
 	/* just guess it's word size */
 	size = dwarf_bytesize(die);
-	if (size <= 0)
+	if ((long)size <= 0)
 		size = default_size;
 
 	return size * 8;
